% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{choose_best_protein_accession}
\alias{choose_best_protein_accession}
\title{Choose the best accession}
\usage{
choose_best_protein_accession(
  input_tbl,
  acc_detail_tab,
  accessions_column,
  row_id_column = uniprot_acc,
  group_id
)
}
\arguments{
\item{input_tbl}{Contain the following columns, 'group_id' which is the Id for each protein group, 'accessions_column' which is the column with the accession of the protein}

\item{accessions_column}{The name of the column with the list of protein accessions, separated by ';' semi-colon. No need to quote the name as we are using tidyverse programming quosure.}

\item{group_id}{The name of the column with the group ID for each protein group. No need to quote the name as we are using tidyverse programming quosure.}

\item{acc_detail_tabl}{The out table from running the function 'parse_fasta_file'}
}
\description{
Choose the best accession
}
