% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{get_plot_num_missing_vales}
\alias{get_plot_num_missing_vales}
\title{Plot the number of missing values in each sample}
\usage{
get_plot_num_missing_vales(input_table)
}
\arguments{
\item{input_table}{Data matrix with each row as a protein and each column a sample.}
}
\value{
A ggplot2 bar plot showing the number of missing values per column.
}
\description{
Plot the number of missing values in each sample
}
