% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoproteomics_helper.R
\name{get_pos_string}
\alias{get_pos_string}
\title{Get PTM position string for the modified peptide}
\usage{
get_pos_string(peptide_start_position, site_relative_position)
}
\arguments{
\item{peptide_start_position:}{Start position of peptide relative to the entire protein sequence}

\item{site_relative_position:}{Position of the modification site relative to the N-terminus of the peptide sequence}
}
\value{
A string. If the position are limited to one unique peptide in the protein, all the phosphosite on that peptide. If the positions are found in multiple repeated peptides in the protein, the phosphosites in each peptide will be contained in round bracket (e.g. (144,148),(170,174),(183,187) ).
}
\description{
get_pos_string
}
