% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{get_significant_data}
\alias{get_significant_data}
\title{Format results table for use in volcano plots, counting number of significant proteins, p-values distribution histogram.}
\usage{
get_significant_data(
  list_of_de_tables,
  list_of_descriptions,
  row_id = uniprot_acc,
  p_value_column = p.mod,
  q_value_column = q.mod,
  log_q_value_column = lqm,
  log_fc_column = logFC,
  comparison_column = comparison,
  expression_column = expression,
  facet_column = analysis_type,
  q_val_thresh = 0.05
)
}
\arguments{
\item{list_of_de_tables}{A list with each element being a results table with log fold-change and q-value per protein.}

\item{list_of_descriptions}{A list of strings describing the parameters used to generate the result table.}

\item{row_id}{The name of the row ID column (tidyverse style).}

\item{p_value_column}{The name of the raw p-value column (tidyverse style).}

\item{q_value_column}{The name of the q-value column (tidyverse style).}

\item{log_q_value_column}{The name of the log q-value column (tidyverse style).}

\item{log_fc_column}{The name of the log fold-change column (tidyverse style).}

\item{comparison_column}{The name of the column describing the contrasts or comparison between groups (tidyverse style).}

\item{expression_column}{The name of the column that will contain the formula expressions of the contrasts.}

\item{facet_column}{The name of the column describing the type of analysis or parameters used to generate the result table (tidyverse style). This is related to the \code{list_of_descriptions} parameter above.}

\item{q_val_thresh}{A numerical value specifying the q-value threshold for statistically significant proteins.}
}
\value{
A table with the following columns:
row_id:  The protein ID, this column is derived from the input to the row_id column.
log_q_value_column: The log (base 10) q-value, this column name is derived from the input to the log_q_value_column.
q_value_column:  The q-value, this column name is derived from the input to the q_value_column.
p_value_column: The p-value, this column name is derived from the input to p_value_column.
log_fc_column: The log fold-change, this column name is derived from the input to log_fc_column.
comparison_column: The comparison, this column name is derived from the input to comparison_column.
expression_column: The formula expression for the contrasts, this column name is derived from the input to expression_column.
facet_column: The analysis type, this column name is derived from the input to facet_column.
colour The colour of the dots used in the volcano plot.
orange = Absolute Log fold-change >= 1 and q-value >= threshold
purple = Absolute Log fold-change >= 1 and q-value < threshold
blue = Absolute Log fold-change < 1 and q-value < threshold
black = all other values
}
\description{
Format results table for use in volcano plots, counting number of significant proteins, p-values distribution histogram.
}
