% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{get_type_of_grouping}
\alias{get_type_of_grouping}
\title{Assign experimental group list}
\usage{
get_type_of_grouping(design_matrix, group_id, sample_id)
}
\arguments{
\item{design_matrix}{A data frame representing the design matrix.}

\item{group_id}{A string representing the name of the group ID column used in the design matrix.}

\item{sample_id}{A string representing the name of the sample ID column used in the design matrix.}
}
\value{
A list where each element name is the name of a treatment group and each element is a vector containing the sample IDs within the treatment group.
}
\description{
Assign experimental group list
}
