% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphoproteomics_helper.R
\name{get_X_mer_string}
\alias{get_X_mer_string}
\title{Get X-mer string given the primary sequence and the position of the PTM}
\usage{
get_X_mer_string(seq, uniprot_acc, position, padding_length = 7)
}
\description{
Given the bioiString object, the uniprot accession, and the position of the phosphorylation site, return the 15mer sequence with the phosphorylation site at the middle.
}
