% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{impute_per_col}
\alias{impute_per_col}
\title{Data imputation function}
\usage{
impute_per_col(temp, width = 0.3, downshift = 1.8)
}
\arguments{
\item{width}{Adjustment factor to the observed standard deviation}

\item{downshift}{Downshift the mean value by this downshift factor multiplied by the observed standard deviation.}

\item{df}{Data matrix}
}
\value{
Data matrix with the missing values from each column replaced with a value randomly sampled from the normal distribution with adjusted mean and standard deviation. The normal distribution parameters are based on the observed distribution of the same column.
}
\description{
Data imputation function
}
