% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_accession_helper.R
\name{parse_fasta_file}
\alias{parse_fasta_file}
\title{Parse the headers of a Uniprot FASTA file and extract the headers and sequences into a data frame
Use seqinr object instead as it seems to be a lot faster to run substring}
\usage{
parse_fasta_file(fasta_file)
}
\arguments{
\item{path}{to input faster file with header format described in https://www.uniprot.org/help/fasta-headers}
}
\value{
A table containing the following columns:
db  sp for Swiss-Prot, tr for TrEMBL
uniprot_acc Uniprot Accession
uniprot_id  Uniprot ID
species     Species
tax_id      Taxonomy ID
gene_name   Gene symbol
protein_evidence 1 to 5, the lower the value, the more evidence that supports the existence of this protein
sequence_version Sequence version
is_isoform  Is it a protein isoform (not the canonical form)
isoform_num     Isoform number.
cleaned_acc Cleaned accession without isoform number.
status  Reviewed or unreviewed.
seq     Amino acid sequence.
seq_length      Sequence length (integer).
}
\description{
Parse the headers of a Uniprot FASTA file and extract the headers and sequences into a data frame
Use seqinr object instead as it seems to be a lot faster to run substring
}
