% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_accession_helper.R
\name{parse_fasta_object}
\alias{parse_fasta_object}
\title{Parse FASTA object from seqinr}
\usage{
parse_fasta_object(aa_seq)
}
\arguments{
\item{aa_seq}{AAStringSet object, output from running seqinr}
}
\value{
A table containing the protein evidence, isoform number, uniprot accession without isoform number, gene name
}
\description{
parse_fasta_object: Parse FASTA headers
}
