% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{print_count_de_genes_table}
\alias{print_count_de_genes_table}
\title{Format results table for use in volcano plots, counting number of significant proteins, p-values distribution histogram.}
\usage{
print_count_de_genes_table(
  list_of_de_tables,
  list_of_descriptions,
  formula_string = "analysis_type ~ comparison",
  facet_column = analysis_type,
  comparison_column = comparison,
  expression_column = expression
)
}
\arguments{
\item{list_of_de_tables}{A list with each element being a results table with log fold-change and q-value per protein.}

\item{list_of_descriptions}{A list of strings describing the parameters used to generate the result table.}

\item{formula_string}{The formula string used in the facet_grid command for the ggplot scatter plot.}

\item{facet_column}{The name of the column describing the type of analysis or parameters used to generate the result table (tidyverse style). This is related to the \code{list_of_descriptions} parameter above.}

\item{comparison_column}{The name of the column describing the contrasts or comparison between groups (tidyverse style).}

\item{expression_column}{The name of the column that will contain the formula expressions of the contrasts.}
}
\description{
Format results table for use in volcano plots, counting number of significant proteins, p-values distribution histogram.
}
