% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{print_p_values_distribution}
\alias{print_p_values_distribution}
\title{Draw the p-values distribution plot.}
\usage{
print_p_values_distribution(
  selected_data,
  p_value_column = p.mod,
  formula_string = "is_ruv_applied ~ comparison"
)
}
\arguments{
\item{selected_data}{A table that is generated by running the function \code{\link{get_significant_data}}.}

\item{formula_string}{The formula string used in the facet_grid command for the ggplot scatter plot.}

\item{log_p_value_column}{The name of the column representing the p-value.}
}
\description{
Draw the p-values distribution plot.
}
