% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{print_volcano_plot}
\alias{print_volcano_plot}
\title{Draw the volcano plot.}
\usage{
print_volcano_plot(
  selected_data,
  log_q_value_column = lqm,
  log_fc_column = logFC,
  q_val_thresh = 0.05,
  formula_string = "analysis_type ~ comparison"
)
}
\arguments{
\item{selected_data}{A table that is generated by running the function \code{\link{get_significant_data}}.}

\item{log_q_value_column}{The name of the column representing the log q-value.}

\item{log_fc_column}{The name of the column representing the log fold-change.}

\item{q_val_thresh}{A numerical value specifying the q-value threshold for statistically significant proteins.}

\item{formula_string}{The formula string used in the facet_grid command for the ggplot scatter plot.}
}
\description{
Draw the volcano plot.
}
