% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove rows in the table where the columns specified by the column regular expression pattern are all zero or NA value.}
\usage{
remove_empty_rows(input_table, col_pattern, row_id)
}
\arguments{
\item{input_table}{Input table with columns recording protein abundances for each sample. The name of these columns matches a regular expression pattern, defined by 'col_pattern'. Remove rows with all samples having no protein abundance.}

\item{col_pattern}{String representing regular expression pattern that matches the name of columns containing the protein abundance values.}

\item{row_id}{The column name with the row_id, tidyverse style name.}
}
\value{
A data frame with the rows without abundance values removed.
}
\description{
Remove rows in the table where the columns specified by the column regular expression pattern are all zero or NA value.
}
