% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{remove_rows_with_missing_values}
\alias{remove_rows_with_missing_values}
\title{For each experimental group, identify proteins that have more than accepted number of missing values per group.}
\usage{
remove_rows_with_missing_values(
  input_table,
  cols,
  design_matrix,
  sample_id,
  row_id,
  group_column,
  max_num_samples_miss_per_group,
  abundance_threshold
)
}
\arguments{
\item{input_table}{An input table with a column containing the row ID and the rest of the columns representing abundance values for each sample.}

\item{cols}{A tidyselect command to select the columns. This includes the functions starts_with(), ends_with(), contains(), matches(), and num_range()}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{sample_id}{The name of the column in design_matrix table that has the sample ID.}

\item{row_id}{A unique ID for each row of the 'input_table' variable.}

\item{group_column}{The name of the column in design_matrix table that has the experimental group.}

\item{max_num_samples_miss_per_group}{An integer representing the maximum number of samples with missing values per group.}

\item{abundance_threshold}{Abundance threshold in which the protein in the sample must be above for it to be considered for inclusion into data analysis.}
}
\value{
A list, the name of each element is the sample ID and each element is a vector containing the protein accessions (e.g. row_id) with enough number of values.
}
\description{
For each experimental group, identify proteins that have more than accepted number of missing values per group.
}
