% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{runTest}
\alias{runTest}
\title{Analyse one contrast (e.g. compare a pair of experimental groups) and output the q-values per protein.}
\usage{
runTest(
  ID,
  A,
  B,
  group_A,
  group_B,
  design_matrix,
  formula_string,
  contrast_variable = "group",
  weights = NA
)
}
\arguments{
\item{ID}{List of protein accessions / row names.}

\item{A}{String representing the name of experimental group A for pairwise comparison of B - A.}

\item{B}{String representing the name of experimental group B for pairwise comparison of B - A.}

\item{group_A}{Names of all the columns / samples that are in experimental group A.}

\item{group_B}{Names of all the columns / samples that are in experimental group B.}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{formula_string}{A formula string representing the experimental design. e.g. ("~ 0 + group")}

\item{contrast_variable}{String representing the contrast variable, which is also used in the formula string. (e.g. "group")}

\item{weights}{Numeric matrix for adjusting each sample and gene.}
}
\value{
A data frame with the following columns:
row.names = the protein accessions
comparison A string showing log({group B's name}) minus log({group A's name})
meanA     mean of the normalized log abundance value of the gene across samples from experimental group A
meanB     mean of the normalized log abundance value of the gene across samples from experimental group B
logFC     log fold-change
tstats    t-test statistics
tmod      moderated t-test statistics
pval      t-test p-value
pmod      moderated t-test p-value
qval      t-test q-value
q.mod     moderated t-test q-value
}
\description{
Analyse one contrast (e.g. compare a pair of experimental groups) and output the q-values per protein.
}
