% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{runTests_contrasts}
\alias{runTests_contrasts}
\title{Run the linear model fitting and statistical tests for a set of contrasts, then adjust with Empirical Bayes function}
\usage{
runTests_contrasts(
  data,
  contrast_strings,
  design_matrix,
  formula_string,
  p_value_column = p.mod,
  q_value_column = q.mod,
  weights = NA
)
}
\arguments{
\item{data}{Data frame containing the log (base 2) protein abundance values where each column represents a sample and each row represents a protein group, and proteins as rows. The data is preferably median-scaled, with missing values imputed, and batch-effects removed.}

\item{contrast_strings}{Input file with a table listing all the experimental contrasts to analyse. It will be in the format required for the function \code{makeContrasts} in the limma package.
The contrast string consists of variable that each consist of concatenating the column name (e.g. group) and the string representing the group type (e.g. A) in the design matrix.}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{formula_string}{A formula string representing the experimental design. e.g. ("~ 0 + group")}

\item{p_value_column}{The name of the raw p-value column (tidyverse style).}

\item{q_value_column}{The name of the q-value column (tidyverse style).}
}
\value{
A list containing two elements. $results returns a list of tables containing logFC and q-values. $fit.eb returns the Empiracle Bayes output object.
}
\description{
Run the linear model fitting and statistical tests for a set of contrasts, then adjust with Empirical Bayes function
}
