% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{go_id_to_term}
\alias{go_id_to_term}
\title{Convert a list of Gene Ontology IDs to their respective human readable name (e.g. GO term).}
\usage{
go_id_to_term(go_string, sep = "; ", goterms, gotypes)
}
\arguments{
\item{goterms}{Output from running \code{goterms <- Term(GOTERM)} from the GO.db library.}

\item{gotypes}{Output from running \code{gotypes <- Ontology(GOTERM)} from the GO.db library.}

\item{A}{string consisting of a list of Gene Ontology ID, separated by a delimiter}
}
\value{
A table with three columns. go_biological_process, go_celluar_compartment, and go_molecular_function. Each column is a list of gene ontology terms, separated by '; '.
}
\description{
Convert a list of Gene Ontology IDs to their respective human readable name (e.g. GO term).
}
