% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_proteins_functions.R
\name{runTests}
\alias{runTests}
\title{Analyse one contrast (e.g. compare a pair of experimental groups) and output the q-values per protein.}
\usage{
runTests(
  ID,
  data,
  tests,
  sample_columns,
  sample_rows_list = NA,
  type_of_grouping,
  design_matrix,
  formula_string,
  contrast_variable = "group",
  weights = NA
)
}
\arguments{
\item{ID}{List of protein accessions / row names.}

\item{design_matrix}{A data frame with a column containing the sample ID (as per the sample_id param) and the experimental group (as per the group param). Each row as the sample ID as row name in the data frame.}

\item{formula_string}{A formula string representing the experimental design. e.g. ("~ 0 + group")}

\item{contrast_variable}{String representing the contrast variable, which is also used in the formula string. (e.g. "group")}

\item{weights}{Numeric matrix for adjusting each sample and gene.}
}
\description{
Analyse one contrast (e.g. compare a pair of experimental groups) and output the q-values per protein.
}
